Rename packages
Rename packages, classes, layouts.
