+++ app/src/main/java/com/example/personalhealth/api/ApiService.kt
@@ -2,7 +2,7 @@ package com.example.personalhealth.api
 
 import com.example.personalhealth.BASE_URL
 import com.example.personalhealth.GET_NUTRIENTS_URL
-import com.example.personalhealth.api.responses.DietListApiResponse
+import com.example.personalhealth.api.responses.RecipesListApiResponse
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
 import retrofit2.http.GET
@@ -14,7 +14,7 @@ private val retrofit = Retrofit.Builder()
 
 interface ApiService{
     @GET(GET_NUTRIENTS_URL)
-    suspend fun getAllNutrients(): DietListApiResponse
+    suspend fun getAllNutrients(): RecipesListApiResponse
 }
 
 object DietsApi{
@@ -22,3 +22,10 @@ object DietsApi{
         retrofit.create(ApiService::class.java)
     }
 }
+
+
+/**
+ * * * * * * * * * * * * * * API * * * * * * * * * * * * * * * * * * *
+ * https://spoonacular.com/food-api/docs#Get-Ingredient-Information *
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ */
