+++ app/src/main/java/com/example/personalhealth/ui/plan/PlanActivity.kt
@@ -1,12 +1,12 @@
-package com.example.personalhealth.ui.aboutus
+package com.example.personalhealth.ui.plan
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.example.personalhealth.R
 
-class AboutUsActivity : AppCompatActivity() {
+class PlanActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_about_us)
+        setContentView(R.layout.activity_plan)
     }
 }
