+++ app/src/main/java/com/example/personalhealth/ui/progress/ProgressActivity.kt
@@ -1,12 +1,12 @@
-package com.example.personalhealth.ui.exercises
+package com.example.personalhealth.ui.progress
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.example.personalhealth.R
 
-class ExerciseActivity : AppCompatActivity() {
+class ProgressActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_exercise)
+        setContentView(R.layout.activity_progress)
     }
 }
