+++ app/src/main/java/com/example/personalhealth/ui/recipes/RecipesViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.personalhealth.ui.diet
+package com.example.personalhealth.ui.recipes
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -7,20 +7,20 @@ import androidx.lifecycle.viewModelScope
 import com.example.personalhealth.Nutrients
 import kotlinx.coroutines.launch
 
-class DietViewModel : ViewModel(){
+class RecipesViewModel : ViewModel(){
 
     private val _nutrientsList = MutableLiveData<List<Nutrients>>()
     val nutrientsList : LiveData<List<Nutrients>>
         get() = _nutrientsList
 
-    private val dietRepository = DietRepository()
+    private val recipesRepository = RecipesRepository()
     init {
         downloadNutrients()
     }
 
     private fun downloadNutrients() {
         viewModelScope.launch {
-            _nutrientsList.value = dietRepository.downloadNutrients()
+            _nutrientsList.value = recipesRepository.downloadNutrients()
         }
     }
 }
