+++ app/src/main/java/com/example/personalhealth/ui/search/SearchActivity.kt
@@ -1,12 +1,12 @@
-package com.example.personalhealth.ui.home
+package com.example.personalhealth.ui.search
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.example.personalhealth.R
 
-class HomeActivity : AppCompatActivity() {
+class SearchActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_home)
+        setContentView(R.layout.activity_search)
     }
 }
