+++ app/src/main/res/layout/activity_plan.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.home.HomeActivity">
+    tools:context=".ui.plan.PlanActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
