+++ app/src/main/res/layout/activity_progress.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.exercises.ExerciseActivity">
+    tools:context=".ui.progress.ProgressActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
