+++ app/src/main/res/layout/activity_recipes.xml
@@ -11,5 +11,5 @@
         android:id="@+id/nutrients_recycler"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.diet.DietActivity"/>
+        tools:context=".ui.recipes.RecipesActivity"/>
 </layout>
